C05_done

+++ C05/ex05/ft_sqrt.c
@@ -1,8 +1,20 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_sqrt.c                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jaewopar <jaewoopk000@naver.com>           +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/10/23 15:49:10 by jaewopar          #+#    #+#             */
+/*   Updated: 2021/10/23 18:44:51 by jaewopar         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 int	ft_sqrt(int nb);
 
 int	ft_sqrt(int nb)
 {
-	unsigned int	n;
+	long	n;
 	int		tmp;
 
 	n = 1;
@@ -11,12 +23,12 @@ int	ft_sqrt(int nb)
 		return (0);
 	else if (nb == 1)
 		return (1);
-	while (n * n <= (unsigned int)nb)
+	while (n * n <= (long)nb)
 	{
-		n++;	
+		n++;
 	}
 	n--;
-	if (n * n == (unsigned int)nb)
-		return (n);
+	if (n * n == (long)nb)
+		return ((int)n);
 	return (0);
 }
